Update Install-LinkPerformance.ps1
Updated Install-LinkPerformance.ps1 to include link to latest version of iPerf. Version used for iPerf in the current release has a TCP Window issue which does not honor OS TCP Scale factor and deals with misleading results. Adding -w and a big TCP window size gets better throughput results with that older version. Not sure if what I modified is enough to get the script working, please review and adapt as necessary.